refactor: centralize search engine configuration
