+++ db/search-engine-client/src/main/java/io/camunda/db/se/config/DatabaseType.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.search.connect.configuration;
+package io.camunda.db.se.config;
 
 public enum DatabaseType {
   ELASTICSEARCH(DatabaseConfig.ELASTICSEARCH),
