+++ db/search-engine-client/src/main/java/io/camunda/db/se/config/SearchEngineConfiguration.java
@@ -1,0 +1,11 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.db.se.config;
+
+public record SearchEngineConfiguration(
+    ConnectConfiguration connect, IndexSettings index, RetentionConfiguration retention) {}
