+++ optimize/util/optimize-commons/pom.xml
@@ -26,6 +26,11 @@
 
     <dependency>
       <groupId>io.camunda</groupId>
+      <artifactId>camunda-search-engine-client</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>io.camunda</groupId>
       <artifactId>camunda-search-client-elasticsearch</artifactId>
     </dependency>
