+++ qa/util/pom.xml
@@ -190,7 +190,7 @@
     </dependency>
     <dependency>
       <groupId>io.camunda</groupId>
-      <artifactId>camunda-search-client-connect</artifactId>
+      <artifactId>camunda-search-engine-client</artifactId>
     </dependency>
 
     <dependency>
