+++ search/search-client-connect/pom.xml
@@ -31,6 +31,11 @@
       <artifactId>camunda-search-client-plugin</artifactId>
     </dependency>
 
+    <dependency>
+      <groupId>io.camunda</groupId>
+      <artifactId>camunda-search-engine-client</artifactId>
+    </dependency>
+
     <!-- elasticsearch -->
 
     <dependency>
