+++ zeebe/exporters/camunda-exporter/src/main/java/io/camunda/exporter/config/ExporterConfiguration.java
@@ -7,9 +7,9 @@
  */
 package io.camunda.exporter.config;
 
-import io.camunda.search.connect.configuration.ConnectConfiguration;
-import java.util.HashMap;
-import java.util.Map;
+import io.camunda.db.se.config.ConnectConfiguration;
+import io.camunda.db.se.config.IndexSettings;
+import io.camunda.db.se.config.RetentionConfiguration;
 
 public class ExporterConfiguration {
 
@@ -115,151 +115,6 @@ public class ExporterConfiguration {
         + ", postExport="
         + postExport
         + '}';
-  }
-
-  public static final class IndexSettings {
-    public static final int DEFAULT_VARIABLE_SIZE_THRESHOLD = 8191;
-    private String prefix = "";
-    private String zeebeIndexPrefix = "zeebe-record";
-
-    private Integer numberOfShards = 1;
-    private Integer numberOfReplicas = 0;
-
-    private Map<String, Integer> replicasByIndexName = new HashMap<>();
-    private Map<String, Integer> shardsByIndexName = new HashMap<>();
-
-    private Integer variableSizeThreshold = DEFAULT_VARIABLE_SIZE_THRESHOLD;
-
-    private boolean shouldWaitForImporters = true;
-
-    public String getPrefix() {
-      return prefix;
-    }
-
-    public void setPrefix(final String prefix) {
-      this.prefix = prefix;
-    }
-
-    public Integer getNumberOfShards() {
-      return numberOfShards;
-    }
-
-    public void setNumberOfShards(final Integer numberOfShards) {
-      this.numberOfShards = numberOfShards;
-    }
-
-    public Integer getNumberOfReplicas() {
-      return numberOfReplicas;
-    }
-
-    public void setNumberOfReplicas(final Integer numberOfReplicas) {
-      this.numberOfReplicas = numberOfReplicas;
-    }
-
-    public Integer getVariableSizeThreshold() {
-      return variableSizeThreshold;
-    }
-
-    public void setVariableSizeThreshold(final Integer variableSizeThreshold) {
-      this.variableSizeThreshold = variableSizeThreshold;
-    }
-
-    public boolean shouldWaitForImporters() {
-      return shouldWaitForImporters;
-    }
-
-    public String getZeebeIndexPrefix() {
-      return zeebeIndexPrefix;
-    }
-
-    public void setZeebeIndexPrefix(final String zeebeIndexPrefix) {
-      this.zeebeIndexPrefix = zeebeIndexPrefix;
-    }
-
-    public void setShouldWaitForImporters(final boolean shouldWaitForImporters) {
-      this.shouldWaitForImporters = shouldWaitForImporters;
-    }
-
-    @Override
-    public String toString() {
-      return "IndexSettings{"
-          + "prefix='"
-          + prefix
-          + '\''
-          + ", numberOfShards="
-          + numberOfShards
-          + ", numberOfReplicas="
-          + numberOfReplicas
-          + ", replicasByIndexName="
-          + replicasByIndexName
-          + ", shardsByIndexName="
-          + shardsByIndexName
-          + ", variableSizeThreshold="
-          + variableSizeThreshold
-          + ", zeebeIndexPrefix='"
-          + zeebeIndexPrefix
-          + '}';
-    }
-
-    public Map<String, Integer> getReplicasByIndexName() {
-      return replicasByIndexName;
-    }
-
-    public void setReplicasByIndexName(final Map<String, Integer> replicasByIndexName) {
-      this.replicasByIndexName = replicasByIndexName;
-    }
-
-    public Map<String, Integer> getShardsByIndexName() {
-      return shardsByIndexName;
-    }
-
-    public void setShardsByIndexName(final Map<String, Integer> shardsByIndexName) {
-      this.shardsByIndexName = shardsByIndexName;
-    }
-  }
-
-  public static final class RetentionConfiguration {
-    private boolean enabled = false;
-    private String minimumAge = "30d";
-    private String policyName;
-
-    public boolean isEnabled() {
-      return enabled;
-    }
-
-    public void setEnabled(final boolean enabled) {
-      this.enabled = enabled;
-    }
-
-    public String getMinimumAge() {
-      return minimumAge;
-    }
-
-    public void setMinimumAge(final String minimumAge) {
-      this.minimumAge = minimumAge;
-    }
-
-    public String getPolicyName() {
-      return policyName;
-    }
-
-    public void setPolicyName(final String policyName) {
-      this.policyName = policyName;
-    }
-
-    @Override
-    public String toString() {
-      return "RetentionConfiguration{"
-          + "enabled="
-          + enabled
-          + ", minimumAge='"
-          + minimumAge
-          + '\''
-          + ", policyName='"
-          + policyName
-          + '\''
-          + '}';
-    }
   }
 
   public static class BulkConfiguration {
