+++ zeebe/exporters/elasticsearch-exporter/pom.xml
@@ -33,6 +33,11 @@
 
     <dependency>
       <groupId>io.camunda</groupId>
+      <artifactId>camunda-search-engine-client</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>io.camunda</groupId>
       <artifactId>camunda-search-client-plugin</artifactId>
       <scope>test</scope>
     </dependency>
